+++ e2e/tests/media.e2e.js
@@ -101,8 +101,7 @@ describe('plugins', () => {
     });
   });
 
-  // eslint-disable-next-line mocha/no-skipped-tests
-  context.skip('full screen', () => {
+  context('full screen', () => {
     before(function() {
       eyesOpen(this);
     });
@@ -116,18 +115,15 @@ describe('plugins', () => {
         cy.get(`[data-hook=${PLUGIN_COMPONENT.IMAGE}]:last`)
           .parent()
           .click();
+        cy.loadOutOfViewImagesInGallery();
+        cy.waitForGalleryImagesToLoad();
         cy.eyesCheckWindow({ tag: this.test.title, target: 'window', fully: false });
       });
     });
 
-    // eslint-disable-next-line mocha/no-skipped-tests
-    context.skip('gallery full screen', () => {
+    context('gallery full screen', () => {
       beforeEach('load editor', () =>
-        cy
-          .loadRicosEditorAndViewer('gallery')
-          .get(`[data-hook=${'image-item'}]:first`)
-          .get(`[data-hook=${'image-item'}]`)
-          .eq(1)
+        cy.loadRicosEditorAndViewer('gallery').waitForGalleryImagesToLoad()
       );
 
       it('expand gallery image on full screen', function() {
@@ -160,8 +156,7 @@ describe('plugins', () => {
     });
   });
 
-  // eslint-disable-next-line mocha/no-skipped-tests
-  context.skip('gallery', () => {
+  context('gallery', () => {
     before(function() {
       eyesOpen(this);
     });
@@ -173,14 +168,12 @@ describe('plugins', () => {
     after(() => cy.eyesClose());
 
     it('render gallery plugin', function() {
-      cy.loadRicosEditorAndViewer('gallery')
-        .get(`[data-hook=${'image-item'}]:first`)
-        .get(`[data-hook=${'image-item'}]`)
-        .eq(1);
+      cy.loadRicosEditorAndViewer('gallery').waitForGalleryImagesToLoad();
       cy.openPluginToolbar(PLUGIN_COMPONENT.GALLERY).shrinkPlugin(PLUGIN_COMPONENT.GALLERY);
       cy.waitForDocumentMutations();
       cy.eyesCheckWindow(this.test.title + ' toolbar');
       cy.openGalleryAdvancedSettings();
+      cy.waitForGalleryImagesToLoad();
       cy.eyesCheckWindow(this.test.title + ' settings');
     });
 
@@ -199,29 +192,30 @@ describe('plugins', () => {
         cy.loadRicosEditorAndViewer('gallery')
           .openPluginToolbar(PLUGIN_COMPONENT.GALLERY)
           .shrinkPlugin(PLUGIN_COMPONENT.GALLERY)
-          .get(`[data-hook=${'image-item'}]:first`)
-          .get(`[data-hook=${'image-item'}]`)
-          .eq(1)
+          .waitForGalleryImagesToLoad()
           .openGalleryAdvancedSettings()
           .openGallerySettings()
-          .get(firstImage)
-          .get(anyImage)
-          .eq(1);
+          .waitForGalleryImagesToLoad();
         cy.eyesCheckWindow(this.test.parent.title + ' - render settings');
         cy.get(firstImage).click();
         cy.get(`[data-hook=${GALLERY_SETTINGS.DELETE}]`);
+        cy.waitForGalleryImagesToLoad();
         cy.eyesCheckWindow(this.test.parent.title + ' - select an item');
         cy.get(`[data-hook=${GALLERY_SETTINGS.SELECT_ALL}]`).click();
+        cy.waitForGalleryImagesToLoad();
         cy.eyesCheckWindow(this.test.parent.title + ' - select all items');
         cy.get(`[data-hook=${GALLERY_SETTINGS.DESELECT}]`).click();
         cy.dragAndDrop(firstImage, anyImage, 1);
+        cy.waitForGalleryImagesToLoad();
         cy.eyesCheckWindow(this.test.parent.title + ' - deselect items');
         cy.get(firstImage).click();
         cy.get(`[data-hook=${GALLERY_SETTINGS.DELETE}]`).click();
         cy.get(firstImage);
+        cy.waitForGalleryImagesToLoad();
         cy.eyesCheckWindow(this.test.parent.title + ' - delete an item');
         cy.get(`[data-hook=${GALLERY_SETTINGS.SELECT_ALL}]`).click();
         cy.get(`[data-hook=${GALLERY_SETTINGS.DELETE}]`).click();
+        cy.waitForGalleryImagesToLoad();
         cy.eyesCheckWindow(this.test.parent.title + ' - delete all items');
       });
     });
@@ -231,9 +225,7 @@ describe('plugins', () => {
         cy.loadRicosEditorAndViewer('gallery')
           .openPluginToolbar(PLUGIN_COMPONENT.GALLERY)
           .shrinkPlugin(PLUGIN_COMPONENT.GALLERY)
-          .get(`[data-hook=${'image-item'}]:first`)
-          .get(`[data-hook=${'image-item'}]`)
-          .eq(1)
+          .waitForGalleryImagesToLoad()
           .openGalleryAdvancedSettings()
           .openGallerySettings()
           .openGalleryImageSettings()
@@ -243,15 +235,19 @@ describe('plugins', () => {
         cy.get(`[data-hook=${GALLERY_IMAGE_SETTINGS.LINK}]`).type('Stunning.com');
         cy.get(`[data-hook=${GALLERY_IMAGE_SETTINGS.LINK_TARGET}]`).click();
         cy.get(`[data-hook=${GALLERY_IMAGE_SETTINGS.LINK_NOFOLLOW}]`).click();
+        cy.waitForGalleryImagesToLoad();
         cy.eyesCheckWindow(this.test.parent.title + ' - enter image settings');
         cy.get(`[data-hook=${GALLERY_IMAGE_SETTINGS.DONE}]:first`).click();
         cy.openGalleryImageSettings();
+        cy.waitForGalleryImagesToLoad();
         cy.eyesCheckWindow(this.test.parent.title + ' - settings saved & title shows on image ');
         cy.get(`[data-hook=${GALLERY_IMAGE_SETTINGS.DELETE}]`).click({ force: true });
         cy.get(`[data-hook=${GALLERY_IMAGE_SETTINGS.PREVIEW}]:first`);
+        cy.waitForGalleryImagesToLoad();
         cy.eyesCheckWindow(this.test.parent.title + ' - delete a media item');
         cy.get(`[data-hook=${GALLERY_IMAGE_SETTINGS.DELETE}]`).click({ force: true });
         cy.get(`[data-hook=${GALLERY_SETTINGS.UPLOAD}]`);
+        cy.waitForGalleryImagesToLoad();
         cy.eyesCheckWindow(this.test.parent.title + ' - delete all items');
       });
       // TODO: title and link image tests