+++ CHANGELOG.md
@@ -19,7 +19,11 @@
 ### :bug: Bug Fix
 - `viewer`
   - [#1557](https://github.com/wix-incubator/rich-content/pull/1557) fix viewer justify css for safari and firefox
- 
+
+### 🏠 Internal
+- `e2e`
+  - [#1555](https://github.com/wix-incubator/rich-content/pull/1555) stabalise flaky gallery tests
+
 </details>
 <hr/>